+++ Assets/YukimaruGames/Terminal/Runtime/Domain/Abstractions.meta
