+++ Assets/YukimaruGames/Terminal/Runtime/Domain/Abstractions/Attributes.meta
