+++ Assets/YukimaruGames/Terminal/Runtime/Domain/Abstractions/Exceptions.meta
