+++ Assets/YukimaruGames/Terminal/Runtime/Domain/Abstractions/Interfaces.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: dad7d1171599c1c4d892a878eff255aa
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
