+++ Assets/YukimaruGames/Terminal/Runtime/Domain/Abstractions/Interfaces/ICommandDiscoverer.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
+using YukimaruGames.Terminal.Domain.Abstractions.Models;
 
-namespace YukimaruGames.Terminal.Domain.API.Commands
+namespace YukimaruGames.Terminal.Domain.Abstractions.Interfaces
 {
     /// <summary>
     /// コマンドを検出するためのインターフェイス.
