+++ Assets/YukimaruGames/Terminal/Runtime/Domain/Abstractions/Interfaces/ICommandHistory.cs.meta
