+++ Assets/YukimaruGames/Terminal/Runtime/Domain/Abstractions/Interfaces/ICommandInvoker.cs
@@ -1,4 +1,6 @@
-namespace YukimaruGames.Terminal.Domain.API.Commands
+using YukimaruGames.Terminal.Domain.Abstractions.Models;
+
+namespace YukimaruGames.Terminal.Domain.Abstractions.Interfaces
 {
     /// <summary>
     /// コマンドの実行インターフェイス.
