+++ Assets/YukimaruGames/Terminal/Runtime/Domain/Abstractions/Interfaces/ICommandParser.cs
@@ -1,6 +1,7 @@
 using System;
+using YukimaruGames.Terminal.Domain.Abstractions.Models;
 
-namespace YukimaruGames.Terminal.Domain.API.Commands
+namespace YukimaruGames.Terminal.Domain.Abstractions.Interfaces
 {
     /// <summary>
     /// コマンド引数のパーサーインターフェイス.
