+++ Assets/YukimaruGames/Terminal/Runtime/Domain/Abstractions/Models/CommandSpecification.cs.meta
