+++ Assets/YukimaruGames/Terminal/Runtime/Domain/Abstractions/YukimaruGames.Terminal.Domain.Abstractions.asmdef
@@ -1,6 +1,6 @@
 {
-    "name": "YukimaruGames.Terminal.Domain.API",
-    "rootNamespace": "YukimaruGames.Terminal.Domain.API",
+    "name": "YukimaruGames.Terminal.Domain.Abstractions",
+    "rootNamespace": "YukimaruGames.Terminal.Domain.Abstractions",
     "references": [
         "GUID:37e9bf61fa8119c4c8e79d1f08c439aa"
     ],
