+++ Assets/YukimaruGames/Terminal/Runtime/Domain/Services/CommandHistory.cs
@@ -1,5 +1,5 @@
 using System.Collections.Generic;
-using YukimaruGames.Terminal.Domain.API.Commands;
+using YukimaruGames.Terminal.Domain.Abstractions.Interfaces;
 
 namespace YukimaruGames.Terminal.Domain.Services
 {
