+++ Assets/YukimaruGames/Terminal/Runtime/Domain/Services/CommandInvoker.cs
@@ -1,4 +1,5 @@
-using YukimaruGames.Terminal.Domain.API.Commands;
+using YukimaruGames.Terminal.Domain.Abstractions.Interfaces;
+using YukimaruGames.Terminal.Domain.Abstractions.Models;
 
 namespace YukimaruGames.Terminal.Domain.Services
 {
