+++ Assets/YukimaruGames/Terminal/Runtime/Domain/Services/CommandLogger.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
-using YukimaruGames.Terminal.Domain.API.Commands;
+using YukimaruGames.Terminal.Domain.Abstractions.Interfaces;
+using YukimaruGames.Terminal.Domain.Abstractions.Models;
 using YukimaruGames.Terminal.SharedKernel;
 
 namespace YukimaruGames.Terminal.Domain.Services
