+++ Assets/YukimaruGames/Terminal/Runtime/Runtime/Context/TerminalRuntimeScope.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 using YukimaruGames.Terminal.Application.Core;
-using YukimaruGames.Terminal.Domain.API.Commands;
+using YukimaruGames.Terminal.Domain.Abstractions.Interfaces;
 
 namespace YukimaruGames.Terminal.Runtime
 {
