+++ Assets/YukimaruGames/Terminal/Runtime/Runtime/Installer/TerminalStandardInstaller.cs
@@ -14,7 +14,7 @@ using System.Collections.Generic;
 using System.Linq;
 using UnityEngine;
 using YukimaruGames.Terminal.Application.Core;
-using YukimaruGames.Terminal.Domain.API.Commands;
+using YukimaruGames.Terminal.Domain.Abstractions.Interfaces;
 using YukimaruGames.Terminal.Domain.Services;
 using YukimaruGames.Terminal.Infrastructure.Commands;
 using YukimaruGames.Terminal.Infrastructure.UI;
